convert all fp64 to fp32

+++ src/pathtracer/bsdf.cu
@@ -27,30 +27,30 @@ DEVICE void make_coord_space(Matrix3x3 &o2w, const Vector3D n) {
 // following code copied from https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#appendix-b-brdf-implementation
 
 // helper: GGX normal distribution
-DEVICE double D_compute(double a, double NoH_raw) {
-  double NoH = fmax(NoH_raw, 0.0); // Heaviside step function
-  double a2 = a*a;
-  double denom = NoH*NoH*(a2 - 1.0) + 1.0;
+DEVICE float D_compute(float a, float NoH_raw) {
+  float NoH = fmaxf(NoH_raw, 0.0); // Heaviside step function
+  float a2 = a*a;
+  float denom = NoH*NoH*(a2 - 1.0) + 1.0;
   return a2 / (PI * denom * denom);
 }
 
-DEVICE double G_compute(
-    double a,    // α = roughness²
-    double NoV,  // max(0, N·V)
-    double NoL,  // max(0, N·L)
-    double VoH,  // dot(V, H)
-    double LoH   // dot(L, H)
+DEVICE float G_compute(
+    float a,    // α = roughness²
+    float NoV,  // max(0, N·V)
+    float NoL,  // max(0, N·L)
+    float VoH,  // dot(V, H)
+    float LoH   // dot(L, H)
 ) {
   // 1) visibility of microfacet only if H·V>0 and H·L>0
   if (VoH <= 0.0 || LoH <= 0.0) return 0.0;
 
   // 2) denominator terms per glTF spec:
   //    G₁(X) = (2·X) / (X + sqrt(α² + (1−α²)·X²))
-  double a2       = a * a;
-  double denomV  = NoV + sqrt(a2 + (1.0 - a2) * NoV * NoV);
-  double denomL  = NoL + sqrt(a2 + (1.0 - a2) * NoL * NoL);
-  double GV      = (2.0 * NoV) / denomV;
-  double GL      = (2.0 * NoL) / denomL;
+  float a2       = a * a;
+  float denomV  = NoV + sqrt(a2 + (1.0 - a2) * NoV * NoV);
+  float denomL  = NoL + sqrt(a2 + (1.0 - a2) * NoL * NoL);
+  float GV      = (2.0 * NoV) / denomV;
+  float GL      = (2.0 * NoL) / denomL;
   return GV * GL;
 }
